temp -> float, post usera bez dzialajacego Date (bo chyba usune to)

+++ new_backend/StationApp/Dtos/StationTemperatureCreateDto.cs
@@ -7,7 +7,7 @@ namespace StationApp.Dtos
     {
         public int TemperatureId { set; get; }
         
-        public int Temperature{ set; get; }
+        public float Temperature{ set; get; }
 
         public long Time { get; set; }
         public DateTime Date{ set; get; }
